\name{SNPQC-package}
\alias{SNPQC-package}
\alias{SNPQC}
\docType{package}
\title{
\packageTitle{SNPQC}
}
\description{
\packageDescription{SNPQC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SNPQC}
\packageIndices{SNPQC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SNPQC}

Maintainer: \packageMaintainer{SNPQC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
